requirements.txt
six
sseclient
synbiochem-py